requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-ai-contentsafety==1.0.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.31.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.0
jiter==0.6.1
lxml==5.3.0
mutagen==1.47.0
openai==1.52.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pypandoc==1.14
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
whisper==1.1.10
yt-dlp==2024.10.7
